refactor(mapgen-studio): extract App.tsx non-React helper corpus into feature modules
Behavior-preserving MOVE of the ~535 LoC of non-React helpers from the top of
App.tsx into focused feature modules. No logic changes: every helper is byte-for-
byte identical, only relocated, and App.tsx re-imports them.

- features/mapConfigSave/api.ts: toConfigId, saveRepoBackedConfig,
  fetchMapConfigSaveDeployStatus, MAP_CONFIG_SAVE_LAST_REQUEST_KEY
- features/runInGame/{api,sourceSnapshotStorage,liveSource}.ts
- features/civ7Setup/{api,setupOptions,livePreset}.ts
- features/configOverrides/configBuilders.ts (merge/path/config builders)
- features/presets/{repoBacked,dialogState}.ts
- shared/{async,number}.ts

localStorage key strings preserved verbatim (persistence contract). tsc clean;
prod build + worker-bundle check pass.
